debug

+++ src/estimator.py
@@ -315,9 +315,11 @@ class Estimator(object):
         print()
         sys.stdout.flush()
 
+        labels_out = labels_out.astype(int)
+
         accuracy = np.mean(np.equal(labels, predicted_labels))
 
-        return activations_out, labels_out, accuracy
+        return activations_out, labels_out.astype(int), accuracy
 
     def robustness(self,
         perturbation_type,
